@@ -1,12 +1,12 @@
     apply plugin: 'com.android.library'
 
     android {
-        compileSdkVersion 24
-        buildToolsVersion '25.0.0'
+        namespace = "umich.cse.yctung.libacousticsensing"
+        compileSdk 33
 
         defaultConfig {
-            minSdkVersion 16
-            targetSdkVersion 24
+            minSdkVersion 31
+            targetSdkVersion 31
             versionCode 1
             versionName "1.0"
 
@@ -18,7 +18,7 @@
 
             ndk {
                 // Note sure why, but x86 will have a bug to compile
-                abiFilters 'armeabi', 'armeabi-v7a', 'arm64-v8a'
+                abiFilters  'armeabi-v7a', 'arm64-v8a'
             }
 
             /*
@@ -45,8 +45,9 @@
     }
 
     dependencies {
-        compile fileTree(include: ['*.jar'], dir: 'libs')
-        testCompile 'junit:junit:4.12'
-        compile 'com.android.support:appcompat-v7:24.0.0'
-        compile 'com.google.code.gson:gson:2.8.0'
+        implementation fileTree(include: ['*.jar'], dir: 'libs')
+        testImplementation 'junit:junit:4.12'
+        androidTestImplementation 'com.google.android:android-test:4.1.1.4'
+        implementation 'androidx.appcompat:appcompat:1.0.0'
+        implementation 'com.google.code.gson:gson:2.8.0'
     }
